+++ landing/src/data/site.ts
@@ -319,5 +319,5 @@ export const site: SiteData = {
 
   analytics: {},
 
-  version: '0.20.0', // x-release-please-version
+  version: '0.21.3', // x-release-please-version
 };